Set up Android java module and package, locally link to sample app
