--- android/src/main/java/com/reactlibrary/RNCavyNativeReporterModule.java
@@ -1,22 +1,0 @@
-
-package com.reactlibrary;
-
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReactContextBaseJavaModule;
-import com.facebook.react.bridge.ReactMethod;
-import com.facebook.react.bridge.Callback;
-
-public class CavyNativeReporterModule extends ReactContextBaseJavaModule {
-
-  private final ReactApplicationContext reactContext;
-
-  public CavyNativeReporterModule(ReactApplicationContext reactContext) {
-    super(reactContext);
-    this.reactContext = reactContext;
-  }
-
-  @Override
-  public String getName() {
-    return "CavyNativeReporter";
-  }
-}
