+++ sampleApp/android/app/build.gradle
@@ -138,6 +138,7 @@ android {
 }
 
 dependencies {
+    implementation project(':cavy-native-reporter')
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "com.android.support:appcompat-v7:${rootProject.ext.supportLibVersion}"
     implementation "com.facebook.react:react-native:+"  // From node_modules
