+++ sampleApp/android/settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'sampleApp'
+include ':cavy-native-reporter'
+project(':cavy-native-reporter').projectDir = new File('../../android')
 
 include ':app'
